requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.0.2
packaging==24.2
pip==24.3.1
pycparser==2.22
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
setuptools==58.0.4
simple-websocket==1.1.0
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
sounddevice==0.5.1